+++ RELEASE-NOTES.md
@@ -6,6 +6,9 @@
 * Removed deprecated methods.
 -->
 
+## Upcoming release
+* Fixed `FormUrlEncoded` body from a dictionary restricted to URL length and not allowing null.
+
 ## 3.3.1
 Released 19 July 2019.